+++ src/App.js
@@ -4,7 +4,7 @@ import {
     Routes,
     Route,
     Link,
-useNavigate,
+    useNavigate,
 } from 'react-router-dom';
 import { PayProvider } from './provider/PayProvider';
 import Home from './pages/Home';
@@ -25,6 +25,9 @@ import Button from './component/Button';
 import NotFoundPage from './pages/404';
 import Test from './pages/Test';
 import PayHistory from './pages/PayHistory';
+import memberPwd from './pages/MemberPwd';
+import CheckPwd from './pages/CheckPwd';
+import MemberPwd from './pages/MemberPwd';
 
 const SplashScreen = () => {
     const navigate = useNavigate();
@@ -65,6 +68,8 @@ function App() {
                         <Route path="/afterkakao" element={<Afterkakao />} />
                         <Route path="*" element={<NotFoundPage />} />
                         <Route path="/history" element={<PayHistory />} />
+                        <Route path="/memberPwd" element={<MemberPwd />} />
+                        <Route path="/checkPwd" element={<CheckPwd />} />
                     </Routes>
                 </PayProvider>
             </Router>
