+++ src/component/BenefitsAndManagement.css
@@ -43,35 +43,12 @@
     text-align: left;
 }
 
-/* 카드 관리 섹션 스타일 */
-.management-section {
-    margin-top: 10px;
-    flex: 1;
-    display: flex;
-    justify-content: space-around;
-}
-
 .management-item {
-    display: flex;
-    flex-direction: column;
     align-items: center;
     justify-content: center;
-    width: 70px;
-    height: 70px;
+    width: 110px;
+    height: 95px;
     background-color: #f0f4f8;
     border-radius: 10px;
     box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
-}
-
-.management-item img {
-    width: 40px;
-    height: 40px;
-    margin-bottom: 5px;
-}
-
-.management-item p {
-    font-size: 12px;
-    color: #333;
-    margin: 0;
-    text-align: center;
 }
