+++ src/pages/Home.css
@@ -1,10 +1,8 @@
-.benefits-section,
-.management-section {
+.benefits-section {
     margin: 20px 0;
 }
 
-.benefits-section h3,
-.management-section h3 {
+.benefits-section h3 {
     text-align: left;
     padding-left: 20px;
     padding-bottom: 20px;
@@ -13,8 +11,7 @@
     margin-bottom: 10px;
 }
 
-.benefits-items,
-.management-items {
+.benefits-items {
     display: flex;
     justify-content: space-around;
     padding: 0 10px;
